README.md: merge history and build info and port to markdown
The adoc files have been merged into the README.md and the
format changed to Markdown because this fork of the repository
is hosted on Github and it is thus easier to read.

Add a section on animation and on-line identification which
is the authors main reason for porting the build to CMake.